122 commits over the May 2026 audit-fixes branch:

Audit phase (commits c217f1cb and earlier, ~50 commits):
- 9 P0 silent-wrong bug fixes (softmax strides, fused LN F64, fill_/scatter_add
  dtype coverage, TENZOR_USE_MKL macro, LN/BN eps cache, cat/slice neg-dim,
  allocator cross-thread, StridedFill complex)
- 10 P1 dtype coverage extensions (creation/math/reductions/scatter/linalg/etc.)
- 3 P1 numerical fixes (norm F64, NaN propagation, F64 Kahan)
- 3 SIMD/threading fixes (runtime dispatch, Int8 wrap, F16 SIMD, false sharing,
  biLSTM MKL gate, OMP consolidation, Philox RNG, determinism mode)
- 6 MKL/oneDNN integration (VML, LAPACKE, gemm_s8u8s32, gemm_bf16bf16f32,
  engine consolidation, format_tag::any)
- 2 allocator improvements (split-sibling coalescing, memory_summary)
- 9 hygiene items (dual LN reg, sparse M/K required, DepthwiseConv per-axis,
  Embedding PaddingIdx, BN storage assert, argmax OMP, OMP threshold env)
- Plus _foreach_* family, tz.special, tz.distributions, __all__ audit

Phase 13 MLIR JIT (commits c217f1cb..0e2b94d5, ~72 commits):
- Text-based MLIR emission + IREE Compiler embedding API + IREE Runtime C API
- 12-op Tenzor dialect via stablehlo.custom_call @tenzor_*
- 4 dialect-op end-to-end plugin path (in-process iree_vm_native_module)
- 50+ ops covered, mini-Llama + ResNet18 forward × CPU + CUDA + Vulkan + ROCm
- Debug UX: show_graph / show_mlir / show_stablehlo / show_iree / cache_stats
- third_party/iree_compat/ (authentic IREE 3.11.0 headers, Apache 2.0)
- .venv-iree/ + third_party/iree_dist/ project-local; FetchContent fallback

158/158 MLIR tests pass.

+++ src/ops/math.cpp
@@ -366,7 +366,7 @@ auto baddbmm(const Tensor& input, const Tensor& batch1, const Tensor& batch2,
     return dispatch_single(OpId::Baddbmm, inputs, attrs);
 }
 
-auto pow(const Tensor& input, float exponent) -> Tensor {
+auto pow(const Tensor& input, double exponent) -> Tensor {
     NewOpAttributes attrs;
     attrs.set(AttrKey::Exponent, static_cast<double>(exponent));
     std::vector<Tensor> inputs = {input};
@@ -398,7 +398,7 @@ auto floor(const Tensor& input) -> Tensor { return detail::unary_op<OpId::Floor>
 auto ceil(const Tensor& input) -> Tensor { return detail::unary_op<OpId::Ceil>(input); }
 auto round(const Tensor& input) -> Tensor { return detail::unary_op<OpId::Round>(input); }
 
-auto clamp(const Tensor& input, float min, float max) -> Tensor {
+auto clamp(const Tensor& input, double min, double max) -> Tensor {
     NewOpAttributes attrs;
     attrs.set(AttrKey::Min, static_cast<double>(min));
     attrs.set(AttrKey::Max, static_cast<double>(max));
@@ -406,14 +406,14 @@ auto clamp(const Tensor& input, float min, float max) -> Tensor {
     return dispatch(OpId::Clamp, inputs, attrs)[0];
 }
 
-auto clamp_min(const Tensor& input, float min) -> Tensor {
+auto clamp_min(const Tensor& input, double min) -> Tensor {
     NewOpAttributes attrs;
     attrs.set(AttrKey::Min, static_cast<double>(min));
     std::vector<Tensor> inputs = {input};
     return dispatch(OpId::ClampMin, inputs, attrs)[0];
 }
 
-auto clamp_max(const Tensor& input, float max) -> Tensor {
+auto clamp_max(const Tensor& input, double max) -> Tensor {
     NewOpAttributes attrs;
     attrs.set(AttrKey::Max, static_cast<double>(max));
     std::vector<Tensor> inputs = {input};
